Refactore code

+++ loja/views.py
@@ -1,20 +1,9 @@
-from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
-from django.shortcuts import render
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import status, generics, pagination
-from rest_framework.decorators import api_view
 from rest_framework.pagination import PageNumberPagination
 from rest_framework.response import Response
-from rest_framework.settings import api_settings
-from rest_framework.throttling import ScopedRateThrottle
-from rest_framework.views import APIView
-from rest_framework import mixins
-from loja.models import *
 from loja.serializers import *
-from rest_framework.pagination import LimitOffsetPagination
-import django_filters.rest_framework
 from rest_framework import filters
-
 from usuario.models import Funcionario
 
 
@@ -57,24 +46,26 @@ class LojaDetalhes(generics.GenericAPIView):
         return Response(lojas_serializer.data)
 
     def put(self, request, id_loja, *args, **kwargs):
-        try:
-            loja = Loja.objects.get(id=id_loja)
-        except Loja.DoesNotExist:
-            return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
-
         if request.user.is_authenticated:
-            funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
-            if funcionario.loja == loja and funcionario.cargo == 'Chefe' or request.user.is_staff:
-                lojas_serializer = LojasSerializer(data=request.data)
-                if lojas_serializer.is_valid():
-                    loja.name = request.data['name']
-                    loja.andar = request.data['andar']
-                    loja.lote = request.data['lote']
-                    loja.cnpj = request.data['cnpj']
-                    loja.save()
-                    return Response(request.data, status=status.HTTP_200_OK)
+            try:
+                loja = Loja.objects.get(id=id_loja)
+            except Loja.DoesNotExist:
+                return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
+
+            if e_funcionario_da_loja(request, loja) or request.user.is_staff:
+                if funcionario_e_chefe(request) or request.user.is_staff:
+                    lojas_serializer = LojasSerializer(data=request.data)
+                    if lojas_serializer.is_valid():
+                        loja.name = request.data['name']
+                        loja.andar = request.data['andar']
+                        loja.lote = request.data['lote']
+                        loja.cnpj = request.data['cnpj']
+                        loja.save()
+                        return Response(request.data, status=status.HTTP_200_OK)
+                    else:
+                        return Response({'erro': 'Json invalid'}, status=status.HTTP_400_BAD_REQUEST)
                 else:
-                    return Response({'erro': 'Json invalid'}, status=status.HTTP_400_BAD_REQUEST)
+                    return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
             else:
                 return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
         else:
@@ -86,10 +77,12 @@ class LojaDetalhes(generics.GenericAPIView):
                 loja = Loja.objects.get(id=id_loja)
             except Loja.DoesNotExist:
                 return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
-            funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
-            if funcionario.loja == loja and funcionario.cargo == 'Chefe' or request.user.is_staff:
-                loja.delete()
-                return Response({'info': 'Item has be deleted'}, status=status.HTTP_204_NO_CONTENT)
+            if e_funcionario_da_loja(request, loja) or request.user.is_staff:
+                if funcionario_e_chefe(request) or request.user.is_staff:
+                    loja.delete()
+                    return Response({'info': 'Item has be deleted'}, status=status.HTTP_204_NO_CONTENT)
+                else:
+                    return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
             else:
                 return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
         else:
@@ -102,7 +95,7 @@ class ProdutosList(generics.GenericAPIView):
 
     filter_backends = [filters.OrderingFilter]
     ordering_fields = ['name', 'id']
-    pagination_class = PageNumberPagination
+
 
     def get(self, request, id_loja, *args, **kwargs):
         try:
@@ -111,7 +104,6 @@ class ProdutosList(generics.GenericAPIView):
             return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
 
         produtos = Produto.objects.filter(loja=id_loja)
-        # produtos_serializer = ProdutoSerializer(produtos, many=True)
         page = self.paginate_queryset(produtos)
         if page is not None:
             serializer = self.get_serializer(page, many=True)
@@ -125,22 +117,18 @@ class ProdutosList(generics.GenericAPIView):
         except Loja.DoesNotExist:
             return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
 
-        if request.user.is_authenticated:
-            funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
-            if funcionario.loja == loja:
-                produto_serializer = ProdutoSerializer(data=request.data)
-                if produto_serializer.is_valid():
-                    name = request.data['name']
-                    quantidade = request.data['quantidade']
-                    valor = request.data['valor']
-                    loja = Loja.objects.get(id=id_loja)
-                    Produto.objects.create(name=name, quantidade=quantidade, valor=valor, loja=loja)
-                    return Response(request.data, status=status.HTTP_201_CREATED)
-                else:
-                    return Response({'erro': 'HTTP_400_BAD_REQUEST / Json is not valid'},
-                                    status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated and e_funcionario_da_loja(request, loja):
+            produto_serializer = ProdutoSerializer(data=request.data)
+            if produto_serializer.is_valid():
+                name = request.data['name']
+                quantidade = request.data['quantidade']
+                valor = request.data['valor']
+                loja = Loja.objects.get(id=id_loja)
+                Produto.objects.create(name=name, quantidade=quantidade, valor=valor, loja=loja)
+                return Response(request.data, status=status.HTTP_201_CREATED)
             else:
-                return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
+                return Response({'erro': 'HTTP_400_BAD_REQUEST / Json is not valid'},
+                                status=status.HTTP_400_BAD_REQUEST)
         else:
             return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
 
@@ -174,18 +162,14 @@ class ProdutoDetalhes(generics.GenericAPIView):
         except Loja.DoesNotExist:
             return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
 
-        if request.user.is_authenticated:
-            funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
-            if funcionario.loja == loja:
-                produto_serializer = ProdutoSerializer(produto, data=request.data)
-                if produto_serializer.is_valid():
-                    produto_serializer.save()
-                    return Response(request.data, status=status.HTTP_200_OK)
-                else:
-                    return Response({'erro': 'HTTP_400_BAD_REQUEST / Json is not valid'},
-                                    status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated and e_funcionario_da_loja(request, loja):
+            produto_serializer = ProdutoSerializer(produto, data=request.data)
+            if produto_serializer.is_valid():
+                produto_serializer.save()
+                return Response(request.data, status=status.HTTP_200_OK)
             else:
-                return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
+                return Response({'erro': 'HTTP_400_BAD_REQUEST / Json is not valid'},
+                                status=status.HTTP_400_BAD_REQUEST)
         else:
             return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
 
@@ -200,13 +184,9 @@ class ProdutoDetalhes(generics.GenericAPIView):
         except Loja.DoesNotExist:
             return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
 
-        if request.user.is_authenticated:
-            funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
-            if funcionario.loja == loja:
-                produto.delete()
-                return Response({'info': 'Item has be deleted'}, status=status.HTTP_204_NO_CONTENT)
-            else:
-                return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
+        if request.user.is_authenticated and e_funcionario_da_loja(request, loja):
+            produto.delete()
+            return Response({'info': 'Item has be deleted'}, status=status.HTTP_204_NO_CONTENT)
         else:
             return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
 
@@ -227,8 +207,7 @@ class ProdutoCompra(generics.GenericAPIView):
             return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
 
         if request.user.is_authenticated:
-            funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
-            if funcionario.loja == loja:
+            if e_funcionario_da_loja(request, loja):
                 dados_serializer = VerifyCompraSerializer(data=request.data)
                 if dados_serializer.is_valid():
                     try:
@@ -267,20 +246,32 @@ class VendaLoja(generics.GenericAPIView):
             loja = Loja.objects.get(id=id_loja)
         except Loja.DoesNotExist:
             return Response({'erro': "HTTP_404_NOT_FOUND"}, status=status.HTTP_404_NOT_FOUND)
-
         if request.user.is_authenticated:
-            funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
-            if funcionario.loja == loja or request.user.is_staff:
+            if e_funcionario_da_loja(request, loja) or request.user.is_staff:
                 vendas = Compra.objects.filter(loja=id_loja)
-
                 page = self.paginate_queryset(vendas)
                 if page is not None:
                     serializer = self.get_serializer(page, many=True)
                     return self.get_paginated_response(serializer.data)
                 serializer = self.get_serializer(vendas, many=True)
                 return Response(serializer.data)
-
             else:
                 return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
         else:
             return Response({'erro': 'HTTP_401_UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
+
+
+def e_funcionario_da_loja(request, loja):
+    if not request.user.is_staff:
+        funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
+        if funcionario.loja == loja:
+            return True
+    return False
+
+
+def funcionario_e_chefe(request):
+    if not request.user.is_staff:
+        funcionario = Funcionario.objects.get(funcionario_complement=request.user.id)
+        if funcionario.cargo == 'Chefe':
+            return True
+    return False
